backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Redis and Celery
redis==5.0.1
celery[redis]==5.3.6

# Config and validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Automation
playwright==1.41.2

# Utilities
python-multipart==0.0.9
email-validator==2.1.0
